project_root/requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
transformers>=4.28.0
numpy>=1.23.5
pandas>=1.5.3
scikit-learn>=1.2.2
pyyaml>=6.0
google-search-results
requests
tweepy
